+++ packages/react-heartwood-components/docz/components/H1.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const H1 = ({ children }) => <h1>{children}</h1>
+
+export default H1