+++ packages/react-heartwood-components/docz/components/index.js
@@ -1,0 +1,2 @@
+export Page from './Page'
+export H1 from './H1'