+++ packages/react-heartwood-components/doczrc.js
@@ -1,0 +1,16 @@
+export default {
+	wrapper: 'docz/DoczWrapper',
+	// theme: 'docz/theme',
+	themeConfig: {
+		colors: {
+			primary: '#1953cb'
+		}
+	},
+	menu: [
+		'React Heartwood Components',
+		{
+			name: 'Buttons',
+			menu: ['Button', 'ButtonGroup', 'ContextMenu']
+		}
+	]
+}