+++ packages/react-heartwood-components/src/components/Page/Page-story.js
@@ -20,7 +20,9 @@ stories
 	.add('Page', () => (
 		<Page
 			header={{
-				title: text('title', '') || 'Page Title'
+				className: 'custom-classname',
+				title: text('title', '') || 'Page Title',
+				collapsed: boolean('isCollapsed', false)
 			}}
 		>
 			<PageContent>
@@ -34,7 +36,8 @@ stories
 				title: text('title', 'Page Title'),
 				backLinkHref: '#',
 				backLinkText: text('backLinkText', 'Back'),
-				hasBottomBorder: boolean('hasBottomBorder', false)
+				hasBottomBorder: boolean('hasBottomBorder', false),
+				collapsed: boolean('isCollapsed', false)
 			}}
 		>
 			<PageContent>
@@ -52,7 +55,8 @@ stories
 					icon: text('icon', ''),
 					isSmall: boolean('isSmall', true)
 				},
-				hasBottomBorder: boolean('hasBottomBorder', false)
+				hasBottomBorder: boolean('hasBottomBorder', false),
+				collapsed: boolean('isCollapsed', false)
 			}}
 		>
 			<PageContent>
@@ -72,7 +76,8 @@ stories
 					icon: text('icon', ''),
 					isSmall: boolean('isSmall', true)
 				},
-				hasBottomBorder: boolean('hasBottomBorder', false)
+				hasBottomBorder: boolean('hasBottomBorder', false),
+				collapsed: boolean('isCollapsed', false)
 			}}
 		>
 			<PageContent>
@@ -86,7 +91,8 @@ stories
 				title: text('title', 'Page Title'),
 				backLinkHref: '#',
 				backLinkText: text('backLinkText', 'Back'),
-				tabs: manyTabs
+				tabs: manyTabs,
+				collapsed: boolean('isCollapsed', false)
 			}}
 		>
 			<PageContent>