+++ packages/spruce-next-helpers/src/skillskit/index.js
@@ -5,6 +5,8 @@ function postMessage(message) {
 	return window.parent.postMessage(JSON.stringify(message), '*')
 }
 
+let skillStatusCheckListener = null
+
 const skill = {
 	editUserProfile: function({
 		userId,
@@ -23,6 +25,17 @@ const skill = {
 		})
 	},
 
+	redirect: function({ route, params = {} }) {
+		Iframes.sendMessage({
+			to: window.parent,
+			eventName: 'Router:Push',
+			data: {
+				route,
+				params
+			}
+		})
+	},
+
 	ready: function({ showHeader = true }: { showHeader: boolean } = {}) {
 		Iframes.sendMessage({
 			to: window.parent,
@@ -32,6 +45,20 @@ const skill = {
 				showHeader
 			}
 		})
+
+		// This is a simple callback that Core can use to see if a skill is
+		// running inside the iframe. Necessary since we don't necessarily want
+		// to allow direct cross-domain access for security reasons.
+		if (!skillStatusCheckListener) {
+			skillStatusCheckListener = Iframes.onMessage(
+				'Skill:StatusCheck',
+				(data, responder) => {
+					responder({
+						eventName: 'Skill:IsAlive'
+					})
+				}
+			)
+		}
 	},
 
 	//TODO move to iframes?
@@ -130,10 +157,11 @@ const skill = {
 					data
 				})
 			},
-			close: () => {
+			close: (data: Object) => {
 				Iframes.sendMessage({
 					to: window.parent,
-					eventName: 'SkillViewDialog:Close'
+					eventName: 'SkillViewDialog:Close',
+					data
 				})
 			},
 			onGoBack: (callback: Function) => {
@@ -162,6 +190,22 @@ const skill = {
 					'SkillViewDialog:ClickFooterSecondaryAction',
 					callback
 				)
+			},
+			onClosed: (callback: Function) => {
+				if (this._onClosedListener) {
+					this._onClosedListener.destroy()
+				}
+				this._onClosedListener = Iframes.onMessage(
+					'SkillViewDialog:Closed',
+					callback
+				)
+			},
+			setContentHeight: (height: string) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'SkillViewDialog:SetContentHeight',
+					data: { value: height }
+				})
 			},
 			setFooterPrimaryActionIsLoading: (isLoading: boolean) => {
 				Iframes.sendMessage({
@@ -222,6 +266,132 @@ const skill = {
 		}
 
 		return modal
+	},
+
+	confirm: function() {
+		const confirm = {
+			show: data => {
+				const { onConfirm, onCancel, ...rest } = data
+
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:Show',
+					data: { ...rest }
+				})
+
+				if (onConfirm) {
+					if (this._onConfirmListener) {
+						this._onConfirmListener.destroy()
+					}
+					this._onConfirmListener = Iframes.onMessage(
+						'Confirm:Confirm',
+						onConfirm
+					)
+				}
+				if (onCancel) {
+					if (this._onCancelListener) {
+						this._onCancelListener.destroy()
+					}
+					this._onCancelListener = Iframes.onMessage('Confirm:Cancel', onCancel)
+				}
+			},
+			setIsConfirming: (isConfirming: boolean) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:SetIsConfirming',
+					data: { value: isConfirming }
+				})
+			},
+			setIsCanceling: (isCanceling: boolean) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:setIsCanceling',
+					data: { value: isCanceling }
+				})
+			},
+			setConfirmButtonIsLoading: (isLoading: boolean) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:SetConfirmButtonIsLoading',
+					data: { value: isLoading }
+				})
+			},
+			setCancelButtonIsLoading: (isLoading: boolean) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:SetCancelButtonIsLoading',
+					data: { value: isLoading }
+				})
+			},
+			setConfirmButtonIsDisabled: (isDisabled: boolean) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:SetConfirmButtonIsDisabled',
+					data: { value: isDisabled }
+				})
+			},
+			setCancelButtonIsDisabled: (isDisabled: boolean) => {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Confirm:SetCancelButtonIsDisabled',
+					data: { value: isDisabled }
+				})
+			}
+		}
+
+		return confirm
+	},
+
+	supportingMessage: function() {
+		const message = {
+			add: function({
+				headline,
+				text,
+				followupText,
+				timeout,
+				kind = 'positive',
+				callback
+			}) {
+				Iframes.sendMessage({
+					to: window.parent,
+					eventName: 'Toast:Add',
+					data: {
+						headline,
+						text,
+						followupText,
+						timeout,
+						kind
+					},
+					onResponse: () => {
+						callback && callback()
+					}
+				})
+			},
+			remove: function() {
+				alert('Not yet implemented...')
+			}
+		}
+
+		return message
+	},
+	blockingMessage: function() {
+		const message = {
+			add: function({
+				headline,
+				text,
+				followupText,
+				timeout,
+				kind = 'positive',
+				callback
+			}) {
+				window.alert(headline)
+			},
+			remove: function() {
+				alert('Not yet implemented...')
+			}
+		}
+
+		return message
 	},
 
 	notifyOfRouteChangeStart() {