+++ packages/spruce-skill-server/services/acl.js
@@ -1,3 +1,4 @@
+const debug = require('debug')('spruce-skill-server')
 // This service can be accessed by ctx.services.acl from the controllers
 
 // From: https://stackoverflow.com/questions/11233498/json-stringify-without-quotes-on-properties
@@ -57,6 +58,18 @@ module.exports = {
 		const { userId, permissions, locationId, organizationId } = options
 		let query
 
+		debug(`Checking ACL permissions`, options)
+
+		if (!userId || !organizationId) {
+			throw new Error(
+				'Missing required parameters "userId" and "organizationId" for ACL check'
+			)
+		}
+
+		if (!permissions) {
+			throw new Error('Missing required parameter "permissions" for ACL check')
+		}
+
 		if (locationId) {
 			query = `
 			{
@@ -91,20 +104,37 @@ module.exports = {
 		}
 		const result = await this.sb.query(query)
 
+		const permsHash = {}
+		Object.keys(permissions).forEach(slug => {
+			permissions[slug].forEach(perm => {
+				permsHash[`${slug}:${perm}`] = false
+			})
+		})
+
 		if (result.data && result.data.Acls) {
 			for (let i = 0; i < result.data.Acls.length; i += 1) {
 				const acl = result.data.Acls[i]
 				if (acl && acl.slug && acl.permissions) {
 					for (let j = 0; j < acl.permissions.length; j += 1) {
 						const permission = acl.permissions[j]
+						permsHash[`${acl.slug}:${permission.name}`] =
+							permission.value === true
 						if (permission.value === false) {
 							return false
 						}
 					}
 				}
 			}
+		} else {
+			debug(`ACL check query failed`, {
+				data: result.data,
+				errors: result.errors
+			})
+			return false
 		}
 
-		return true
+		const isAllowed = Object.values(permsHash).every(value => value === true)
+
+		return isAllowed
 	}
 }