+++ packages/spruce-skill/server/tests/ExampleTests.js
@@ -1,16 +1,16 @@
 // @flow
 const { assert } = require('chai')
-// const faker = require('faker')
-// const { Op } = require('sequelize')
-const Base = require('./Base')
+const { SpruceTest } = require('@sprucelabs/spruce-skill-server')
 
-class ExampleTests extends Base {
+// SpruceTest take a single parameter, pointing to the base skill directory
+class ExampleTests extends SpruceTest(`${__dirname}/../../`) {
 	organization: any
 	location: any
 
 	setup() {
 		it('Can do a trivial assert', () => this.trivialAssert())
 		it('Can get users', () => this.getUsers())
+		it('Can use custom mock data', () => this.customMock())
 	}
 
 	async before() {
@@ -46,6 +46,16 @@ class ExampleTests extends Base {
 
 		log.debug(body)
 		assert.isNotNull(body.data.Users)
+	}
+
+	async customMock() {
+		// Verify "someData" is set properly by mocks/ExampleMock.js
+		assert.equal(
+			this.mocks.example.someData,
+			`Example Mock Test org/location: ${this.organization.id} / ${
+				this.location.id
+			}`
+		)
 	}
 }