+++ packages/eslint-plugin-spruce/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.15.2](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.15.1...v8.15.2) (2019-06-19)
+
+**Note:** Version bump only for package eslint-plugin-spruce
+
+
+
+
+
+## [8.15.1](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.15.0...v8.15.1) (2019-06-19)
+
+**Note:** Version bump only for package eslint-plugin-spruce
+
+
+
+
+
+# [8.15.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.14.0...v8.15.0) (2019-06-18)
+
+**Note:** Version bump only for package eslint-plugin-spruce
+
+
+
+
+
+# [8.14.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.13.0...v8.14.0) (2019-06-14)
+
+
+### Bug Fixes
+
+* **lint:** Update dependencies for shared lint config ([89b255f](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/89b255f))
+
+
+### Features
+
+* **lint:** Enforce GQL linting in skills ([b9d9b00](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/b9d9b00))
+* **lint:** Enforce that query/mutate must be called with an object of options. ([78d2c3f](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/78d2c3f))
+
+
+
+
+
 # [8.13.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v8.12.0...v8.13.0) (2019-03-20)
 
 **Note:** Version bump only for package @sprucelabs/eslint-plugin-spruce